Move histogram expressions to a different directory
