+++ src/renderer/viz/expressions.js
@@ -557,7 +557,10 @@ import ViewportMin from './expressions/aggregation/viewport/ViewportMin';
 import ViewportSum from './expressions/aggregation/viewport/ViewportSum';
 import ViewportCount from './expressions/aggregation/viewport/ViewportCount';
 import ViewportPercentile from './expressions/aggregation/viewport/ViewportPercentile';
-import ViewportHistogram from './expressions/aggregation/viewport/ViewportHistogram';
+// import ViewportHistogram from './expressions/aggregation/viewport/ViewportHistogram';
+import ViewportHistogram from './expressions/histogram/ViewportHistogram';
+import GlobalHistogram from './expressions/histogram/GlobalHistogram';
+import ViewportFeatures from './expressions/viewportFeatures';
 
 import GlobalAvg from './expressions/aggregation/global/GlobalAvg';
 import GlobalCount from './expressions/aggregation/global/GlobalCount';
@@ -565,8 +568,6 @@ import GlobalMax from './expressions/aggregation/global/GlobalMax';
 import GlobalMin from './expressions/aggregation/global/GlobalMin';
 import GlobalSum from './expressions/aggregation/global/GlobalSum';
 import GlobalPercentile from './expressions/aggregation/global/GlobalPercentile';
-
-import ViewportFeatures from './expressions/viewportFeatures';
 
 import Zoom from './expressions/zoom';
