+++ test/integration/user/test/viewportHistogram.test.js
@@ -2,7 +2,7 @@ import carto from '../../../../src/index';
 import * as util from '../../util';
 import { OTHERS_INDEX } from '../../../../src/renderer/viz/expressions/constants';
 
-describe('viewportHistogram', () => {
+fdescribe('viewportHistogram', () => {
     let div, map, source, viz, layer;
 
     // Some useful features
